+++ packages/core/src/components/accordion/components/accordion-section.interfaces.ts
@@ -1,4 +1,4 @@
-export type AccordionSectionState = "success" | "info" | "warning" | "error";
+export type AccordionSectionState = "success" | "info" | "warning" | "danger" | "error";
 
 export type AccordionHeading = "h2" | "h3" | "h4" | "h5";
 
@@ -6,5 +6,6 @@ export const stateMap: Record<AccordionSectionState, string> = {
   success: "succes:",
   info: "info:",
   warning: "waarschuwing:",
+  danger: "fout:",
   error: "fout:",
 };