+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## NEXT
 
-### Changed
-* **BREAKING:** Accordion: status danger eruit (deprecate), status error erin ([#1976](https://github.com/dso-toolkit/dso-toolkit/issues/1976))
+### Added
+* Accordion: status error toegevoegd ([#1976](https://github.com/dso-toolkit/dso-toolkit/issues/1976))
+
+### Deprecated
+* Accordion: status danger is deprecated. Gebruik voortaan status warning of status error ([#1976](https://github.com/dso-toolkit/dso-toolkit/issues/1976))
 
 ## 50.0.0